--- src/wazuh_testing/constants/executions.py
@@ -1,17 +1,0 @@
-'''
-copyright: Copyright (C) 2015-2023, Wazuh Inc.
-           Created by Wazuh, Inc. <info@wazuh.com>.
-           This program is free software; you can redistribute it and/or modify it under the terms of GPLv2
-'''
-import pytest
-
-
-# Marks Executions
-
-TIER0 = pytest.mark.tier(level=0)
-TIER1 = pytest.mark.tier(level=1)
-TIER2 = pytest.mark.tier(level=2)
-
-AGENT = pytest.mark.agent
-SERVER = pytest.mark.server
-LINUX = pytest.mark.linux
