+++ src/wazuh_testing/utils/configuration.py
@@ -411,55 +411,3 @@ def update_feed_path_configurations(configurations, metadata, feeds_path):
                                                    os.path.join(feeds_path, 'nvd', metadata[index]['nvd_feed'])))
 
     return new_configurations
-
-
-def set_correct_prefix(configurations, new_prefix):
-    """Insert the correct prefix in the paths of each configuration.
-
-    In Mac OS X it is not possible to create files in the / directory.
-    Therefore, it is necessary to replace those paths that do not contain a
-    suitable prefix.
-
-    This function checks if the path inside directories, ignore, nodiff and restrict sections
-    contains a certain prefix, and if it does not contain it, it inserts it.
-
-    Args:
-        configurations (list): List of configurations loaded from the YAML.
-        new_prefix (str): Prefix to be inserted before every path.
-
-    Returns:
-        configurations (list): List of configurations with the correct prefix added in the directories and
-        ignore sections.
-    """
-
-def load_wazuh_configurations(yaml_file_path: str, test_name: str, params: list = None, metadata: list = None) -> object():
-    r"""
-    Load different configurations of Wazuh from a YAML file.
-
-    Args:
-        yaml_file_path (str): Full path of the YAML file to be loaded.
-        test_name (str): Name of the file which contains the test that will be executed.
-        params (list, optional) : List of dicts where each dict represents a replacement
-        MATCH/REPLACEMENT. Default `None`
-        metadata (list, optional): Custom metadata to be inserted in the configuration. Default `None`
-
-    Returns:
-        Python object with the YAML file content
-    Raises:
-        ValueError: If the length of `params` and `metadata` are not equal.
-    """
-    params = [{}] if params is None else params
-    metadata = [{}] if metadata is None else metadata
-    if len(params) != len(metadata):
-        raise ValueError(f"params and metadata should have the same length {len(params)} != {len(metadata)}")
-
-    with open(yaml_file_path) as stream:
-        configurations = yaml.safe_load(stream)
-
-    if sys.platform == 'darwin':
-        configurations = set_correct_prefix(configurations, ROOT_PREFIX)
-
-    return [process_configuration(configuration, placeholders=replacement, metadata=meta)
-            for replacement, meta in zip(params, metadata)
-            for configuration in configurations
-            if test_name in expand_placeholders(configuration.get('apply_to_modules'), placeholders=replacement)]
