+++ src/wazuh_testing/utils/db_queries/global_db.py
@@ -83,22 +83,6 @@ def clean_agents_from_db():
         raise Exception('Unable to clean agents')
 
 
-
-def insert_agent_in_db(id=1, name='TestAgent', ip='any', registration_time=0, connection_status=0,
-                       disconnection_time=0):
-    """
-    Write agent in global.db
-    """
-    insert_command = f'global insert-agent {{"id":{id},"name":"{name}","ip":"{ip}","date_add":{registration_time}}}'
-    update_command = f'global sql UPDATE agent SET connection_status = "{connection_status}",\
-                       disconnection_time = "{disconnection_time}" WHERE id = {id};'
-    try:
-        database.query_wdb(insert_command)
-        database.query_wdb(update_command)
-    except Exception:
-        raise Exception(f"Unable to add agent {id}")
-
-
 # Insert agents into DB and assign them into a group
 def insert_agent_into_group(total_agents):
     for i in range(total_agents):
@@ -113,16 +97,6 @@ def insert_agent_into_group(total_agents):
                    "source":"remote","data":[{{"id":{id},"groups":["Test_group{id}"]}}]}}'''
         results = database.query_wdb(command)
         assert results == 'ok'
-
-
-def remove_db_agent(agent_id):
-    """Function that wraps the needed queries to remove an agent.
-
-    Args:
-        agent_id(int): Unique identifier of an agent
-    """
-    data = database.query_wdb(f"global delete-agent {agent_id}").split()
-    assert data[0] == 'ok', f"Unable to remove agent {agent_id} - {data[1]}"
 
 
 def calculate_global_hash():
@@ -160,10 +134,3 @@ def clean_belongs():
         database.query_wdb(command)
     except Exception:
         raise Exception('Unable to clean belongs table.')
-
-
-def list_agents_ids():
-    wazuhdb_result = database.query_wdb('global get-all-agents last_id -1')
-    list_agents = [agent['id'] for agent in wazuhdb_result if not (0 == agent.get('id'))]
-
-    return list_agents
