+++ src/wazuh_testing/utils/time.py
@@ -8,6 +8,7 @@ import subprocess
 import sys
 import threading
 import time
+import numbers
 from datetime import datetime, timedelta
 
 
@@ -296,3 +297,12 @@ def parse_date_time_format(date_time: str):
             return f"{match.group(1)} {match.group(2)}{item['append']}"
 
     ValueError(f"Could not parse the {date_time} datetime.")
+
+def validate_interval_format(interval):
+    """Validate that the interval passed has the format in which the last digit is a letter from those passed and
+       the other characters are between 0-9."""
+    if interval == '':
+        return False
+    if interval[-1] not in ['s', 'm', 'h', 'd', 'w', 'y'] or not isinstance(int(interval[0:-1]), numbers.Number):
+        return False
+    return True
